minor rework to move collecting files into a reusable bit of API so that we can use it also for building the feed pretty easily with the last N
